Show the recipes
Here is included the 3 differents recipes that I made. The process to build the pages was easy because I already praticed the whole workflow before

The only thing I missed once was the <img attribute as "scr" instead of "src", but I quickly noticed it and changed it. After that, the rest of the recipes was just a copy of the code from a recipes and just changes the informations in it, so it was fast