Center .button
I was struggling with centering the button element on the pages before. I entered all necessaries except the one that helps apply all the changes to the natural behaviors to the element <a> which contains the .button class. Naturally <a> is an inline element, but since I have written it not inline another element, which is usually the case that I've praticed, I missed its normal display behavior. so, I have added the display property to _block_ to make sure the width and the margin could be respected properly and now it works.